move room calls to gameservice
